+++ tests/testthat/test-notin.R
@@ -1,0 +1,3 @@
+test_that("notin works", {
+  expect_equal(c("a", "d") %notin% c("a", "b"), c(FALSE, TRUE))
+})